Hotfixes

+++ contracts/tokenSwapping/IExchange.sol
@@ -3,7 +3,7 @@
 pragma solidity ^0.8.0;
 
 interface IExchange {
-    function deposit(address tokenAddress, uint amount) external;
+    function deposit(address tokenAddress, uint256 amount) external;
 
-    function withdraw(address tokenAddress, uint amount) external;
+    function withdraw(address tokenAddress, uint256 amount) external;
 }
